Added refresh token logic for 401

+++ app/src/main/java/de/westnordost/streetcomplete/data/workspace/domain/model/LoginResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LoginResponse(
     val access_token: String,
-    val expires_in: Int,
-    val refresh_expires_in: Int,
+    val expires_in: Long,
+    val refresh_expires_in: Long,
     val refresh_token: String
 )
